More vulkan stuff in the window object, python-free reference build.
The VulkanWindow struct now also creates swapchain image views, command pools,
some fence and semaphore stuff, and some teardown code for all of that.

For some reason, calling vkAcquireNextImageKHR when running in cppyy causes an
access violation somewhere inside vkAcquireNextImageKHR.

I added a clang-only build to compare with, and it does not have this problem.